[SER-447] WIP move string to rsharp crate.
