+++ client/rust/src/contract/properties/mod.rs
@@ -242,24 +242,7 @@ pub(crate) use primitives;
 
 #[cfg(feature = "c_api")]
 pub mod c_api {
-    /// Implements From traits to convert references of wrapper types using `ref-cast`
-    macro_rules! impl_from_refcast {
-        ($from_t:ty, $for_t:ty) => {
-            impl<'a> From<&'a $from_t> for &'a $for_t {
-                fn from(other: &'a $from_t) -> &'a $for_t {
-                    use ref_cast::RefCast;
-                    <$for_t>::ref_cast(other)
-                }
-            }
-            impl<'a> From<&'a mut $from_t> for &'a mut $for_t {
-                fn from(other: &'a mut $from_t) -> &'a mut $for_t {
-                    use ref_cast::RefCast;
-                    <$for_t>::ref_cast_mut(other)
-                }
-            }
-        };
-    }
-    pub(crate) use impl_from_refcast;
+    pub(crate) use rsharp::impl_from_refcast;
 
     pub mod c_types {
         pub use crate::contract::c_api::ContractDataHandle;
@@ -272,25 +255,10 @@ pub mod c_api {
         pub use i32;
         pub use i64;
         pub use i8;
+        pub use rsharp::c_api::_string;
         pub use u16;
         pub use u32;
         pub use u64;
         pub use u8;
-
-        pub use super::_string::String;
-    }
-
-    mod _string {
-        use derive_more::{From, Into};
-        use safer_ffi::prelude::*;
-
-        #[derive_ReprC]
-        #[ReprC::opaque]
-        #[repr(transparent)]
-        #[derive(ref_cast::RefCast, From, Into, Debug, Clone, Eq, PartialEq, Hash)]
-        pub struct String {
-            inner: std::string::String,
-        }
-        impl_from_refcast!(std::string::String, String);
     }
 }
